Small change to heartbeat message sending logic
Now it waits for an extra loop after ``maxSend`` just in order to give the
remotes time to reply to the last round of messages, but id does not send an
extra message.

If no strange filtering or throttling is going on, the number of received
messages from any main remote should be ``maxSend`` whenever we fail to
receive from other remotes.